Move HTML validation from naucse.utils.views to naucse.validation
The validator wasn't even used in views!
